correctly lock password on '--lock-user'
If the documentation for manpages of chpasswd [1] and usermod [2] is
correct then this change removes the need for jaunty and karmic suites to
override the update_passwords from dapper.   The copied code had resulted
in jaunty and karmic suites not locking the user account if '--lock-user'
was passed.

I've tested building a jaunty image with 
  ./vmbuilder xen ubuntu --suite jaunty --arch i386 \
      --dest xen-jaunty-i386 --part part --debug --lock-user
  ./vmbuilder kvm ubuntu --suite hardy --arch i386 -\
      --dest xen-hardy-i386 --part part --debug --lock-user

and verified that the user's account is password locked for password
login, but not expired (ie, has a '!' at the beginning of the encrypted
password entry, but no expiration value).  Also, verified that after
removing the '!' at the beginning I was able to login.

example entry in /etc/shadow:
  jaunty:  ubuntu:!$6$<trimmed>/:14497:0:99999:7:::
  hardy: ubuntu:$1$<trimmed>:14497:0:99999:7:::

log output shows use of 'usermod -L' for 'Locking' in both, and 
'chpasswd --md5' in hardy and 'chpasswd' in jaunty.

To fix bug LP#420581, automated-ec2-builds/build-ec2-image must also be
changed to pass '--lock-user' to vmbuilder.

Signed-off-by: Scott Moser <smoser@ubuntu.com>

--
[1] http://manpages.ubuntu.com/cgi-bin/search.py?ie=UTF-8&q=chpasswd&titles=Title&lr=lang_en
[2] http://manpages.ubuntu.com/cgi-bin/search.py?ie=UTF-8&q=usermod&titles=Title&lr=lang_en