Sometimes, trying to write a unit test for something makes you realise how incredibly silly it is.
Hence:

Replace NonBufferedFile with NonBlockingFile.

NonBufferedFile was, in fact, buffered. It used various tricks to pretend to
be non-blocking.

NonBlockingFile is, in fact, also buffered, but actually non-blocking,
so this implementation is much cleaner (and faster, but this is not really
performance sensitive code anyways).

Refactor run_cmd to use this new NonBlockingFile.